separate main air
